+++ R/ustreasuries.R
@@ -10,6 +10,7 @@
 #' @section Vignettes:
 #' \itemize{
 #'     \item \bold{veronesi-ch02} Examples and figures from Veronesi, Ch2
+#'     \item \bold{veronesi-ch02-Nelson-Seigel} Appendix to Veronesi, Ch2 showing term-structure modeling
 #'     \item \bold{cmt-rates} a description of Constant Maturity and
 #'     Annualized Percentage rates
 #'     \item \bold{yield-curves} examples of downloading the data and printing
@@ -63,6 +64,7 @@
 #'                  \item \bold{Nelson.Siegel.plot} Plot the actual and Nelson.Siegel Yield Curves
 #'                  \item \bold{NSzeros.plot} Plot Z(0, T) and r(0, T)
 #'                  \item \bold{Zbootstrap} Derive Z(0, T_i) using the bootstrap method
+#'                  \item \bold{spot_rate} Derive spot rate from Z(0, T) and T
 #'              }
 #'      }
 #'      \itemize{